Modified default behavior of ssu-prep. Previously, if none of -x, -p, -s, --fp, --fs were enabled then the program would create a script that would run all jobs serially (not in parallel), which requires the user to manually edit the output shell script to make it actually work in parallel (by adding qsub prefix for example).
I decided the default mode should produce a script that runs in
parallel, so now ssu-prep requires a 4th command line varialbe, the
prefix/suffix file. The -p, -s, --fp, --fs options (which controlled
the definition of prefix and suffix) are now gone. The -x option
changes the usage so only 3 arguments are expected on the cmdline
(prefix/suffix file is not expected). With -x, the jobs are run in
parallel on an n-core machine. A new option, -y, specifies that also
the 4th command line variable is not expected, and the resulting
script will run in serial mode, thus requiring a user to edit it to
get the jobs to run in parallel (thus -y now reproduces the default
mode of the previous behavior).



git-svn-id: https://svn.janelia.org/eddylab/nawrockie/src@980 2857e7bf-25ef-0310-8806-9752169b59e1